Windows high contrast mode changes
Using the (unsupported) -ms-high-contrast media query to tidy up the
homepage.

These changes are based on feedback by a visually impaired user using
high contrast mode on a Windows computer.

The person noted that where the hero image is, there is just blank space
 (high contrast mode doesn't seem to display background images) so the
 user wasn't sure if they were missing something. Therefore I suggest
 removing the background image and arrow icon in high contrast mode.

With the 'promo' boxes, it's functionality is that the whole box is a
link. When hovering over the area with a mouse, the cursor changes to a
pointer as expected. Hovering over bits where there is no text (blank
space) similarly raised the question of whether the user was missing
something. Therefore I suggest adding a border around the promo to show
the area the box covers. This can be done by adding a 1px transparent
border, which obviously has no affect when not in high contrast mode.